pyproject.toml
[tool.poetry]
name = "habits-course-work"
version = "0.1.0"
description = ""
authors = ["DalOpez <alex-grazhdankin@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
psycopg2-binary = "^2.9.10"
djangorestframework-simplejwt = "^5.5.0"
python-dotenv = "^1.1.1"
black = "^25.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
